requirements.txt
django==5.1.6
python-dotenv
psycopg2-binary
social-auth-app-django
djangorestframework
whitenoise
gunicorn